# openapi.yaml
openapi: 3.0.3
info:
  title: EventSpin API
  description: |
    API para la plataforma EventSpin de ruletas de premios. Permite gestionar eventos,
    colecciones, ítems, spins, canjes y uploads de imágenes.

    **Autenticación**: La mayoría de los endpoints requieren un JWT Bearer token en el
    header `Authorization: Bearer <token>`. Los tokens se obtienen mediante `/auth/login`.

    **Rate limiting**:
    - Auth endpoints: 10 req/min
    - Spin endpoint: 60 req/min
    - Resto: 120 req/min
  version: 1.0.0
  contact:
    email: guidogauna9@gmail.com
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Unauthorized
...
    ItemSnapshot:
      type: object
      properties:
        name:
          type: string
        imageURL:
          type: string
          format: uri
        metadata:
          type: object
          additionalProperties: true
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 1
      responses:
        '200':
          description: Login exitoso
          headers:
            Set-Cookie:
              description: refresh_token httpOnly cookie (7 días)
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                        description: JWT (15 min)
                      user:
                        type: object
                        properties:
                          id:
                            type: string
                          email:
                            type: string
                          role:
                            type: string
                            enum: [org_admin, super_admin, operator]
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Invalid credentials
        '403':
          description: Cuenta no verificada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Demasiadas solicitudes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar access token
      description: Usa el cookie `refresh_token` httpOnly para emitir un nuevo access token y rotar el refresh token.
      responses:
        '200':
          description: Nuevo access token emitido
          headers:
            Set-Cookie:
              description: Nuevo refresh_token cookie rotado
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
        '401':
          description: Refresh token ausente, expirado o revocado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /events/{eventId}/stats:
    get:
      tags: [Stats]
      summary: Estadísticas del evento
      description: |
        Soporta dos modos:
        - **JSON** (default): Snapshot de estadísticas actuales.
        - **SSE** (`Accept: text/event-stream`): Stream en tiempo real con actualizaciones cada 5 segundos.
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
        - name: Accept
          in: header
          schema:
            type: string
            enum: [application/json, text/event-stream]
      responses:
        '200':
          description: Estadísticas del evento (JSON o SSE)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      event:
                        type: object
                        properties:
                          id:
                            type: string
                          name:
                            type: string
                          status:
                            type: string
                          slug:
                            type: string
                      kpis:
                        type: object
                        properties:
                          totalSpins:
                            type: integer
                          totalWinners:
                            type: integer
                          pending:
                            type: integer
                          redeemed:
                            type: integer
                          expired:
                            type: integer
                          totalItems:
                            type: integer
                          totalCollections:
                            type: integer
                          spinsLast24h:
                            type: integer
                      lastWinners:
                        type: array
                        items:
                          type: object
                          properties:
                            code:
                              type: string
                            itemSnapshot:
                              $ref: '#/components/schemas/ItemSnapshot'
                            redeemedAt:
                              type: string
                              format: date-time
                      itemsDistribution:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                            count:
                              type: integer
                      winRate:
                        type: string
                        example: '20.0'
                        description: Porcentaje de spins ganadores
            text/event-stream:
              schema:
                type: string
                description: Stream SSE con eventos JSON del mismo esquema cada 5 segundos

  # ─── COLLECTIONS ─────────────────────────────────────────────────────────────
...